+++ web/packages/api/src/index.ts
@@ -135,4 +135,5 @@ export * as toPolkadot from './toPolkadot'
 export * as toEthereum from './toEthereum'
 export * as utils from './utils'
 export * as status from './status'
+export * as assets from './assets'
 export * as environment from './environment'
