+++ web/packages/api/src/utils.ts
@@ -30,7 +30,7 @@ export const paraIdToChannelId = (paraId: number): string => {
 
 export const forwardedTopicId = (messageId: string): string => {
     // From rust code
-	// (b"forward_id_for", original_id).using_encoded(sp_io::hashing::blake2_256)
+    // (b"forward_id_for", original_id).using_encoded(sp_io::hashing::blake2_256)
     const typeEncoded = stringToU8a('forward_id_for')
     const paraIdEncoded = hexToU8a(messageId)
     const joined = new Uint8Array([...typeEncoded, ...paraIdEncoded])
